clean up vyper_compile tests
